web/requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
structlog==23.2.0
supabase==2.3.0
